requirements.txt
streamlit
python-dotenv
openai-agents
langsmith[openai-agents]
openai
requests
chromadb
